microservices/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pyjwt==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
requests==2.31.0
